# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/messages": {
      "post": {
        "operationId": "MessagesController_sendMessage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Messages"
        ]
      }
    },
...
    "/messages/{friendId}": {
      "get": {
        "operationId": "MessagesController_getMessages",
        "parameters": [
          {
            "name": "friendId",
            "required": true,
            "in": "path",
            "description": "Id do usuário",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Messages"
        ]
      }
    }
  },
  "info": {
    "title": "Heat Me Up API",
    "description": "API para videochat com amigos e desconhecidos.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateMessageDto": {
        "type": "object",
        "properties": {
          "toUserId": {
            "type": "string",
            "description": "Id do usuário receptor da mensagem",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "content": {
            "type": "string",
            "description": "Conteúdo da mensagem",
            "example": "Olá, tudo bem?"
          }
        },
        "required": [
          "toUserId",
          "content"
        ]
      }
    }
  }
}